fix(web): move UserAccount inside store provider
To persist the state even on the refresh, we have to store the state in localstorage and again we have to retrieve the data from storage and update the store. But this will lead to error if not properly updated. This is fixed by setting the state in UserAccount component and that component is moved inside store provider.